Optimizing the receiving functions for speed.
The recieve functions seem to be the major bottleneck, always waiting a
long time before pushing things back at the sender. Jerry rigged a
function that sends stuff back as soon as it recieves it and it's twice
as fast, but it can still be much improved. Must also make it more
universal.