7x increase in speed and clean up of repo.
Changed script to not use Queues for threading, improved the thread
structure to cut the waiting time in half. This has results in a 7x
times speed increase. Used cProfile to verify this. The next steps are
to clean up the main threading function to make it easier to modify and
then finally create a udpProxy class based on the tcpProxy class.
Additionally, cleaned up the repository a bit using the gitignore file.
Should have done that earlier.