Began udpProxy class, tweaks to ClientThread
Tweaked the ClientThread function to work nicely with tcpProxy and
udpProxy classes. Tests show it works great for TCP, however more tweaks
must be made for udpProxy. Thinking about creating a socket wrapper for
ClientThread to interact with. Debating if overhead might cause
performance issues and whether not to write a seperate threading
subclass for udpProxy.