Almost working model, some bugs
The receiving proxy has troubles with how to handle the incoming
fragmented packets because a new thread is started for each fragment.
Thinking of a way to only start a thread if a full packet has been
recieved and some mechanism to clean out this buffer if fragments have
been in their too long.